Split out the mixins and replaced containment checks with matchers
 * Moved the mixins to separate files to make them easier to read
 * Moved out containment matches to individual matcher objects
 * Added much more intelligent out of order matching
