+++ src/h_matchers/exception.py
@@ -1,0 +1,5 @@
+"""Exceptions for use in h-matchers."""
+
+
+class NoMatch(ValueError):
+    """The items being compared do not match."""
