+++ src/h_matchers/matcher/collection/_mixin/__init__.py
@@ -1,0 +1,6 @@
+"""Mixins to extend the capabilities of the AnyCollection matcher."""
+
+from h_matchers.matcher.collection._mixin.contains import ContainsMixin
+from h_matchers.matcher.collection._mixin.item_matcher import ItemMatcherMixin
+from h_matchers.matcher.collection._mixin.size import SizeMixin
+from h_matchers.matcher.collection._mixin.type import TypeMixin
